Hide editor toolbar when no field is focused
- I could imagine this is to be annoying when working with the fields
  because it causes a layout shift
- On the other hand, especially in the browser, the extra space would be
  nice
