--- ts/editor/Fields.svelte
@@ -1,23 +1,0 @@
-<!--
-Copyright: Ankitects Pty Ltd and contributors
-License: GNU AGPL, version 3 or later; http://www.gnu.org/licenses/agpl.html
--->
-<div class="fields">
-    <slot />
-</div>
-
-<style lang="scss">
-    .fields {
-        display: grid;
-        grid-auto-rows: min-content;
-        grid-gap: 4px;
-
-        /* moves the scrollbar inside the editor */
-        overflow-x: hidden;
-
-        > :global(:last-child) {
-            /* bottom padding is eaten by overflow-x */
-            margin-bottom: 5px;
-        }
-    }
-</style>
